Resolve some merge conficts and restored some GPU code functionality. Reworked HSA Locale Model. The test case loopy.c should now run correctly The merged reduction code in compiler/AST/build.cpp has not been fixed and may still need to be commented out for this to run.

+++ runtime/include/chplcgfns.h
@@ -87,7 +87,7 @@ extern chpl_fn_p chpl_ftable[];
 
 #ifdef TARGET_HSA
 extern const char* chpl_gpu_kernels[];
-extern const int64_t chpl_num_gpu_kernels;
+extern const int32_t chpl_num_gpu_kernels;
 #endif
 extern chpl_fn_info chpl_finfo[];
